% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_flux_qaqc}
\alias{run_flux_qaqc}
\title{Launch the interactive QA/QC app for flux data}
\usage{
run_flux_qaqc(offset)
}
\arguments{
\item{offset}{Integer. The site's UTC offset in hours (e.g. –5 for “UTC–5”, +2 for “UTC+2”).
This argument is required.}
}
\description{
Launch the interactive QA/QC app for flux data
}
